refactor: split contract into domain modules
